+++ src/plugins/common/components/seo-fields/index.jsx
@@ -64,7 +64,7 @@ const SeoFields = ({ areLinkedImagesAllowed, isEditMode, node }) => {
           isEditMode={isEditMode}
           node={node}
           label="SEO Image"
-          modalTitle="Select SEO Image"
+          title="Select SEO Image"
         />
         <Field
           component={DatePickerField}