+++ src/plugins/dam/components/image-asset-picker-field/index.jsx
@@ -27,7 +27,7 @@ class ImageAssetPickerField extends React.Component {
     isEditMode: PropTypes.bool,
     label: PropTypes.string,
     meta: PropTypes.object.isRequired, // eslint-disable-line react/forbid-prop-types
-    modalTitle: PropTypes.string,
+    title: PropTypes.string,
     node: PropTypes.instanceOf(Message),
   };
 
@@ -35,7 +35,7 @@ class ImageAssetPickerField extends React.Component {
     areLinkedImagesAllowed: true,
     isEditMode: true,
     label: '',
-    modalTitle: '',
+    title: '',
     node: null,
   };
 
@@ -84,7 +84,7 @@ class ImageAssetPickerField extends React.Component {
       isEditMode,
       label,
       meta: { error },
-      modalTitle,
+      title,
       node,
       getNode,
     } = this.props;
@@ -155,7 +155,7 @@ class ImageAssetPickerField extends React.Component {
           areLinkedImagesAllowed={areLinkedImagesAllowed}
           assetTypes={[imageType]}
           isOpen={isModalOpen}
-          modalTitle={modalTitle}
+          title={title}
           node={node}
           onCloseUploader={this.handleCloseUploader}
           onSelectImage={this.handleSelectImage}