+++ src/plugins/dam/components/image-asset-picker-modal/index.jsx
@@ -29,7 +29,7 @@ class ImageAssetPickerModal extends React.Component {
     galleryNode: PropTypes.instanceOf(Message),
     isOpen: PropTypes.bool,
     label: PropTypes.string,
-    modalTitle: PropTypes.string,
+    title: PropTypes.string,
     multiAssetErrorMessage: PropTypes.string,
     node: PropTypes.instanceOf(Message),
     onCloseUploader: PropTypes.func,
@@ -44,7 +44,7 @@ class ImageAssetPickerModal extends React.Component {
     galleryNode: null,
     isOpen: false,
     label: 'Select Primary Image',
-    modalTitle: '',
+    title: '',
     multiAssetErrorMessage: 'Invalid Action: Trying to assign multiple files as a Primary Media.',
     node: null,
     onCloseUploader: noop,
@@ -115,7 +115,7 @@ class ImageAssetPickerModal extends React.Component {
       galleryNode,
       isOpen,
       label,
-      modalTitle,
+      title,
       multiAssetErrorMessage,
       node,
       onSelectImage,
@@ -128,7 +128,7 @@ class ImageAssetPickerModal extends React.Component {
       <div>
         <Modal onOpened={this.handleOpened} centered size="xxl" isOpen={isOpen} toggle={onToggleModal}>
           <ModalHeader toggle={onToggleModal}>
-            <span className="nowrap">{modalTitle !== '' ? modalTitle : label}</span>
+            <span className="nowrap">{title !== '' ? title : label}</span>
           </ModalHeader>
           <ModalBody className="p-0">
             {