+++ src/plugins/ovp/components/video-fields/index.jsx
@@ -141,7 +141,7 @@ const VideoFields = ({ formName, isEditMode, schemas, mezzanine, video }) => (
           component={ImageAssetPickerField}
           isEditMode={isEditMode}
           label="Poster Image"
-          modalTitle="Select Poster Image"
+          title="Select Poster Image"
           name="posterImageRef"
           node={video}
         />